\name{mvmstan-package}
\alias{mvmstan-package}
\alias{mvmstan}
\docType{package}
\title{
Simplifies specification of (multivariate) mixture models in Stan (mc-stan.org).
~~ mvmstan ~~
}
\description{
mvmstan lets you specify a graphical model of a mixture and translates it into Stan code.
}
\details{
\tabular{ll}{
Package: \tab mvmstan\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-01-14\cr
License: \tab CC1.0\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
